checkpoint, fixing type errors, rename api directory
