+++ api/lib/middlewares/with-error-response.ts
