+++ api/static-routes.ts
