+++ tsconfig.json
@@ -29,13 +29,6 @@
     // "rootDir": "./",                                  /* Specify the root folder within your source files. */
     "moduleResolution": "Bundler" /* Specify how TypeScript looks up a file from a given module specifier. */,
     "baseUrl": "./" /* Specify the base directory to resolve non-relative module names. */,
-    "paths": {
-      "@api/*": ["./dev-server-api/*"],
-      "@cli/*": ["./cli/*"],
-      "@frontend/*": ["./dev-server-frontend/*"],
-      // deprecated: use @api/* instead
-      "@server/*": ["./dev-server-api/src/*"],
-    } /* Specify a set of entries that re-map imports to additional lookup locations. */,
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
     // "types": [],                                      /* Specify type package names to be included without being referenced in a source file. */
@@ -126,7 +119,4 @@
   //   "dist",
   //   "**/dist"
   // ]
-  "exclude": [
-    "**/routes/**/*.ts"
-  ]
 }
